requirements.txt
# Web framework
fastapi[all]

# DB section
sqlalchemy
alembic

# GRPC section
grpcio
grpcio-tools

# Quality gates and testing section
pycodestyle

# graphQL section
strawberry-graphql[fastapi]
py-graphql-client
```

